Aufgabe 5/PubSub/Pub_sub function/requirements.txt
# Function dependencies, for example:
# package>=version
datetime
google-cloud-storage
google-cloud-pubsub
requests
pandas